# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  devops-cd_internal_dto.ApplicationResponse:
    properties:
      app_type:
        type: string
      created_at:
        type: string
      default_depends_on:
        items:
          type: integer
        type: array
      deployed_tag:
        description: 当前部署的镜像标签
        type: string
      description:
        type: string
      env_clusters:
        additionalProperties:
          items:
            type: string
          type: array
        description: 环境集群配置，从 app_env_configs 表查询得出
        type: object
      id:
        type: integer
      name:
        type: string
      project_id:
        description: 关联的项目ID
        type: integer
      project_name:
        description: 项目名称
        type: string
      repo_id:
        type: integer
      repo_name:
        description: Repository的namespace/name
        type: string
      status:
        type: integer
      team_id:
        type: integer
      team_name:
        description: 团队名称
        type: string
      updated_at:
        type: string
    type: object
...
  devops-cd_internal_dto.CreateRepositoryRequest:
    properties:
      description:
        type: string
      git_type:
        enum:
        - gitea
        - gitlab
        - github
        type: string
      git_url:
        type: string
      language:
        type: string
      name:
        maxLength: 100
        type: string
      namespace:
        maxLength: 100
        type: string
      project_id:
        type: integer
      team_id:
        type: integer
    required:
    - git_type
    - git_url
    - name
    - namespace
    type: object
  devops-cd_internal_dto.CreateTeamRequest:
    properties:
      description:
        type: string
      leader_name:
        maxLength: 100
        type: string
      name:
        maxLength: 100
        type: string
      project_id:
        type: integer
    required:
    - name
    - project_id
    type: object
...
  devops-cd_internal_dto.DeleteRepositoryRequest:
    properties:
      id:
        description: 必填：要删除的代码库ID
        type: integer
    required:
    - id
    type: object
...
  devops-cd_internal_dto.PageResponse:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  devops-cd_internal_dto.RepositoryResponse:
    properties:
      applications:
        description: 关联的应用列表
        items:
          $ref: '#/definitions/devops-cd_internal_dto.ApplicationResponse'
        type: array
      created_at:
        type: string
      description:
        type: string
      full_name:
        description: namespace/name
        type: string
      git_type:
        type: string
      git_url:
        type: string
      id:
        type: integer
      language:
        type: string
      name:
        type: string
      namespace:
        type: string
      project_id:
        type: integer
      project_name:
        type: string
      status:
        type: integer
      team_id:
        type: integer
      team_name:
        type: string
      updated_at:
        type: string
    type: object
...
  devops-cd_internal_dto.TeamResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      leader_name:
        type: string
      name:
        type: string
      project_id:
        type: integer
      updated_at:
        type: string
    type: object
  devops-cd_internal_dto.TeamSimpleResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      project_id:
        type: integer
    type: object
...
  devops-cd_internal_dto.UpdateRepositoryRequest:
    properties:
      description:
        type: string
      git_type:
        enum:
        - gitea
        - gitlab
        - github
        type: string
      git_url:
        type: string
      id:
        description: 必填：要更新的代码库ID
        type: integer
      language:
        type: string
      name:
        maxLength: 100
        type: string
      namespace:
        maxLength: 100
        type: string
      project_id:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
      team_id:
        type: integer
    required:
    - id
    type: object
  devops-cd_internal_dto.UpdateTeamRequest:
    properties:
      description:
        type: string
      id:
        type: integer
      leader_name:
        maxLength: 100
        type: string
      name:
        maxLength: 100
        type: string
      project_id:
        type: integer
    required:
    - id
    type: object
...
  devops-cd_pkg_utils.Response:
    properties:
      code:
        type: integer
      data: {}
      detail:
        description: 详细错误信息（可选）
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    DevOps 持续交付平台 API 文档
    提供代码库管理、应用管理、批次管理、部署管理等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: DevOps CD API
  version: "1.0"
paths:
...
  /api/v1/repositories:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 项目名称
        in: query
        name: project
        type: string
      - description: 团队ID
        in: query
        name: team_id
        type: integer
      - description: Git类型
        in: query
        name: git_type
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 状态
        in: query
        name: status
        type: integer
      - description: 是否包含应用列表
        in: query
        name: with_applications
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.PageResponse'
              type: object
      summary: 获取代码库列表
      tags:
      - Repository
  /api/v1/repository:
    get:
      consumes:
      - application/json
      parameters:
      - description: 代码库ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.RepositoryResponse'
              type: object
      summary: 获取代码库详情
      tags:
      - Repository
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建代码库请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.CreateRepositoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.RepositoryResponse'
              type: object
      summary: 创建代码库
      tags:
      - Repository
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新代码库请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.UpdateRepositoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.RepositoryResponse'
              type: object
      summary: 更新代码库
      tags:
      - Repository
  /api/v1/repository/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 删除代码库请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.DeleteRepositoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/devops-cd_pkg_utils.Response'
      summary: 删除代码库
      tags:
      - Repository
  /api/v1/team:
    get:
      consumes:
      - application/json
      parameters:
      - description: 团队ID
        format: int64
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.TeamResponse'
              type: object
      summary: 获取团队详情
      tags:
      - Team
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建团队请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.TeamResponse'
              type: object
      summary: 创建团队
      tags:
      - Team
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新团队请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.UpdateTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.TeamResponse'
              type: object
      summary: 更新团队
      tags:
      - Team
  /api/v1/team/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 团队ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/devops-cd_pkg_utils.Response'
      summary: 删除团队
      tags:
      - Team
  /api/v1/teams:
    get:
      consumes:
      - application/json
      parameters:
      - description: 项目ID
        in: query
        name: project_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/devops-cd_internal_dto.TeamSimpleResponse'
                  type: array
              type: object
      summary: 获取团队列表（返回所有团队，可按项目过滤，用于下拉选择）
      tags:
      - Team
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"